requirements.txt
asgiref==3.2.7
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.9
pycodestyle==2.6.0
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
